feat(proposals): add PythLazerProposals and PythCoreProposals components
